pdf_ocr_chat/requirements.txt
flask
python-dotenv
google-cloud-vision
PyPDF2
pdf2image
pillow
groq
flask-socketio